+++ Source/PluginProcessor.h
@@ -159,6 +159,7 @@ public:
     void getLowEndMeterLevels(float& subRms, float& lowStereoRisk, float& outputPeak) const noexcept;
     void getPumpMeterLevels(float& phase, float& gain, float& reduction, bool& active) const noexcept;
     void getGuardMeterLevels(float& drive, float& reduction, bool& active) const noexcept;
+    void requestFxSendTailKill() noexcept;
     struct HostSyncStatus
     {
         double bpm = 124.0;